requirements/requirements-conda.txt
brax
chex
distrax
evox @ git+https://github.com/ZaberKo/evox@main
flashbax
gymnax
jaxmarl
jumanji
optax
orbax-checkpoint
pgx
